test file and other softening options
