+++ src/watertap_contrib/seto/unit_models/chemical_softening_0D.py
@@ -8,8 +8,12 @@ from pyomo.environ import (
     check_optimal_termination,
     Param,
     Suffix,
+    value,
     log,
     units as pyunits,
+    Expression,
+    NonNegativeReals
+    
 )
 from pyomo.common.config import ConfigBlock, ConfigValue, In
 
@@ -34,6 +38,7 @@ from watertap.core import InitializationMixin
 __author__ = "Kurban Sitterley, Abdiel Lugo"
 
 _log = idaeslog.getLogger(__name__)
+
 
 
 class SofteningTrainType(StrEnum):
@@ -137,8 +142,6 @@ class ChemicalSoftening0DData(InitializationMixin, UnitModelBlockData):
 
         ### REFERENCES ###
 
-        ion_set = self.config.property_package.ion_set
-
         # this creates blank scaling factors, which are populated later
         self.scaling_factor = Suffix(direction=Suffix.EXPORT)
 
@@ -170,32 +173,43 @@ class ChemicalSoftening0DData(InitializationMixin, UnitModelBlockData):
         prop_in = self.properties_in[0]
         prop_out = self.properties_out[0]
         prop_waste = self.properties_waste[0]
-        comps = self.config.property_package.component_list
-
-        if "TSS" in comps:
-            self.frac_TSS_removal = Var(
-                initialize=0.85, doc="Default 85% removal of TSS"
-            )
+        comps = self.config.property_package.component_set       
+        non_hardness_comps = [j for j in self.config.property_package.component_set if j not in ["Ca_2+","Mg_2+"]]
 
 
-        self.lime_mw = Param(
-            initialize=74, units=pyunits.g / pyunits.mol, doc="Molecular weight of lime"
-        )
 
         self.ca_eff_target = Var(
-            initialize=0.030, units=pyunits.kg / pyunits.m**3
+            initialize=0.030, 
+            units=pyunits.kg / pyunits.m**3,
+            doc = "Target Ca in CaCO3 equivalents in the effluent"
         )
 
         self.mg_eff_target = Var(
-            initialize=0.020, units=pyunits.kg / pyunits.m**3
+            initialize=0.020, 
+            units=pyunits.kg / pyunits.m**3,
+            doc = "Target Mg in CaCO3 equivalents in the effluent"
         )
 
-        self.frac_vol_recovery = Var(
-            initialize=0.99, units=pyunits.dimensionless, 
-            doc="Fractional volumetric recovery of water"
+
+        # Creating default effluent composition of all solutes other than Ca and Mg
+
+        removal_eff_dict = dict(
+            zip(
+                non_hardness_comps,
+                [
+                    0.7 if j != "TDS" else 1e-3
+                    for j in non_hardness_comps
+                ],
+            )
         )
 
-        self.excess_lime = Var(initialize=100, units=pyunits.mg / pyunits.liter)
+        self.removal_efficiency = Var(
+            non_hardness_comps,
+            initialize=removal_eff_dict,
+            doc="Removal efficiency of all solutes other than Ca and Mg",
+        )
+
+        # System design variables
 
         self.retention_time_mixer = Var(
             initialize=1,
@@ -232,20 +246,40 @@ class ChemicalSoftening0DData(InitializationMixin, UnitModelBlockData):
             doc="Sedimentation basin overflow rate",
         )
 
+        self.no_of_mixer = Var(
+            initialize=1, 
+            units= pyunits.dimensionless, 
+            doc="Volume of mixer"
+        )
+
+        self.no_of_floc = Var(
+            initialize=1, 
+            units= pyunits.dimensionless, 
+            doc="Volume of mixer"
+        )
+
         self.volume_mixer = Var(
-            initialize=10, units=pyunits.m**3, doc="Volume of mixer"
+            initialize=10, 
+            units=pyunits.m**3, 
+            doc="Volume of mixer"
         )
 
         self.volume_floc = Var(
-            initialize=10, units=pyunits.m**3, doc="Volume of flocculator"
+            initialize=10, 
+            units=pyunits.m**3, 
+            doc="Volume of flocculator"
         )
 
         self.volume_sed = Var(
-            initialize=10, units=pyunits.m**3, doc="Volume of sedimentation basin"
+            initialize=10, 
+            units=pyunits.m**3, 
+            doc="Volume of sedimentation basin"
         )
 
         self.volume_recarb = Var(
-            initialize=10, units=pyunits.m**3, doc="Volume of recarbonator"
+            initialize=10, 
+            units=pyunits.m**3, 
+            doc="Volume of recarbonator"
         )
 
         self.vel_gradient_mix = Var(
@@ -262,167 +296,457 @@ class ChemicalSoftening0DData(InitializationMixin, UnitModelBlockData):
             doc="Velocity gradient of flocculator",
         )
 
-        self.vel_gradient_sed = Var(
-            initialize=15,
-            units=pyunits.s**-1,
-            bounds=(10, 50),
-            doc="Velocity gradient of sedimentation",  # ???
+        self.Ca_CaCO3_conv = Param(
+            initialize = 2.5,
+            units = pyunits.dimensionless,
+            doc = "Conversion factor for Ca to equivalent CaCO3"
         )
+
+        self.Mg_CaCO3_conv = Param(
+            initialize = 4.12,
+            units = pyunits.dimensionless,
+            doc = "Conversion factor for Ca to equivalent CaCO3"
+        )
+
+        self.frac_vol_recovery = Var(
+            initialize=0.99, 
+            units=pyunits.dimensionless, 
+            bounds = (0,None),
+            doc="Fractional volumetric recovery of water"
+        )
+
+        # Create softening chemical dosing variables based on softening model
+
+        if self.config.softening_procedure_type is SofteningProcedureType.single_stage_lime:
+
+            self.CaOH_dosing = Var(
+            initialize = 1e5,
+            units = pyunits.kg / pyunits.day,
+            bounds = (0,None),
+            doc = "Lime requirements"
+            )
+        
+            self.CO2_first_basin = Var(
+            initialize = 1e2,
+            units = pyunits.kg / pyunits.day,
+            bounds = (0,None),
+            doc = "CO2 concentration required for recarbonation with single basin"
+            )
+
+        elif self.config.softening_procedure_type is SofteningProcedureType.excess_lime:
+
+            self.excess_CaOH = Var(
+            initialize=100, 
+            units= pyunits.kg / pyunits.m**3,
+            doc= "Excess lime requiremenent"
+            )
+
+            self.CaOH_dosing = Var(
+            initialize = 1e5,
+            units = pyunits.kg / pyunits.day,
+            bounds = (0,None),
+            doc = "Lime requirements"
+            )
+        
+            self.CO2_first_basin = Var(
+            initialize = 1e2,
+            units = pyunits.kg / pyunits.day,
+            bounds = (0,None),
+            doc = "CO2 concentration required for recarbonation with single basin"
+            )
+
+        elif self.config.softening_procedure_type is SofteningProcedureType.single_stage_lime_soda:
+
+            self.CaOH_dosing = Var(
+            initialize = 1e5,
+            units = pyunits.kg / pyunits.day,
+            bounds = (0,None),
+            doc = "Lime requirements"
+            )
+            
+            self.Na2CO3_dosing = Var(
+            initialize = 1e5,
+            units = pyunits.kg / pyunits.day,
+            bounds = (0,None),
+            doc = "Soda ash requirements"
+            )
+
+            self.CO2_first_basin = Var(
+            initialize = 1e2,
+            units = pyunits.kg / pyunits.day,
+            bounds = (0,None),
+            doc = "CO2 concentration required for recarbonation with single basin"
+             )
+
+        else:
+
+            self.CaOH_dosing = Var(
+                initialize = 1e5,
+                units = pyunits.kg / pyunits.day,
+                bounds = (0,None),
+                doc = "Lime requirements"
+            )        
+
+            self.Na2CO3_dosing = Var(
+                initialize = 1e5,
+                units = pyunits.kg / pyunits.day,
+                bounds = (0,None),
+                doc = "Soda ash requirements"
+            )
+
+            self.CO2_first_basin = Var(
+                initialize = 1e2,
+                units = pyunits.kg / pyunits.day,
+                bounds = (0,None),
+                doc = "CO2 concentration required for recarbonation with single basin"
+            )
+
+            self.CO2_second_basin = Var(
+                initialize = 1e2,
+                units = pyunits.kg / pyunits.day,
+                bounds = (0,None),
+                doc = "CO2 concentration required for recarbonation in second basin only in excess lime scenario"
+            )
+
+
+        self.excess_CaOH = Var(
+            initialize=0, 
+            units= pyunits.kg / pyunits.m**3,
+            doc= "Excess lime requiremenent"
+            )        
+        
+        self.CO2_CaCO3 = Var(
+            initialize = 115,
+            units = pyunits.kg / pyunits.m**3,
+            bounds = (0,None),
+            doc = "Carbonic acid in mg/L as CaCO3"
+        )
+
+        self.MgCl2_dosing = Var(
+            initialize = 0,
+            units = pyunits.kg / pyunits.m**3,
+            bounds = (0,None),
+            doc = "Magnesium Chloride requirments dosing"
+        )
+
+        self.sludge_prod  = Var(
+            initialize = 1e6,
+            units = pyunits.kg / pyunits.day,
+            bounds = (0,None),
+            doc= "Sludge production rate in kg/day"
+        )
+
+        # Add Ca,Mg carbonate hardness --> Created as temporary variables so that basic property can be used
+        if "Ca_2+" in comps:    
+            @self.Expression(doc="Calcium in influent converted to equivalent CaCO3")
+            def Ca_CaCO3(b):
+                return (pyunits.convert(prop_in.conc_mass_comp["Ca_2+"]*b.Ca_CaCO3_conv, to_units= pyunits.kg / pyunits.m**3))
+        else: 
+            self.Ca_CaCO3 = Param(
+            initialize = 0,
+            units = pyunits.kg / pyunits.m**3,
+            doc="Calcium in influent converted to kg/m3 as CaCO3"
+            )
+ 
+
+        if "Mg_2+" in comps:
+            @self.Expression(doc="Magnesium in influent converted to kg/m3 as CaCO3")
+            def Mg_CaCO3(b):
+                return (pyunits.convert(prop_in.conc_mass_comp["Mg_2+"]*b.Mg_CaCO3_conv, to_units= pyunits.kg / pyunits.m**3))
+        else:
+            self.Mg_CaCO3 = Param(
+            initialize = 0,
+            units = pyunits.kg / pyunits.m**3,
+            doc="Magnesium in influent converted to kg/m3 as CaCO3"
+            )
+
+
+        if 'Alkalinity_2-' in comps:
+            self.check_alk = Expression( expr= self.Ca_CaCO3 - prop_in.conc_mass_comp["Alkalinity_2-"])
+
+            if self.check_alk is NonNegativeReals:
+                self.Ca_hardness_CaCO3 = Expression(expr = prop_in.conc_mass_comp["Alkalinity_2-"])
+                self.Ca_hardness_nonCaCO3 = Expression( expr= self.Ca_CaCO3 - prop_in.conc_mass_comp["Alkalinity_2-"])
+                self.Mg_hardness_CaCO3 = Expression( expr= 0)
+                self.Mg_hardness_nonCaCO3 = Expression( expr = self.Mg_CaCO3 )
+            else:
+                self.Ca_hardness_CaCO3 = Expression( expr = self.Ca_CaCO3 )
+                self.Ca_hardness_nonCaCO3 = Expression( expr= 0)
+                self.Mg_hardness_CaCO3 =  Expression( expr= prop_in.conc_mass_comp['Alkalinity_2-'] - self.Ca_CaCO3 )
+                self.Mg_hardness_nonCaCO3 = Expression( expr= self.Ca_CaCO3 + self.Mg_CaCO3 - prop_in.conc_mass_comp['Alkalinity_2-'])
+
+
+        # Calculating chemical dosing
+
+        if self.config.softening_procedure_type is SofteningProcedureType.single_stage_lime:
+
+            @self.Constraint(doc="Lime dosing")
+            def eq_CaOH_dosing(b):
+                return b.CaOH_dosing == pyunits.convert((b.CO2_CaCO3 + b.Ca_hardness_CaCO3) * 0.56 * prop_in.flow_vol,
+                                                         to_units=pyunits.kg/pyunits.d)
+                                                           
+            @self.Constraint(doc="CO2 for first basin")
+            def eq_CO2_first_basin(b): 
+                return b.CO2_first_basin == pyunits.convert((prop_in.conc_mass_comp["Alkalinity_2-"] - b.Ca_CaCO3 
+                                                + prop_out.conc_mass_comp["Ca_2+"]* b.Ca_CaCO3_conv) * prop_in.flow_vol,to_units = pyunits.kg/pyunits.d) * 0.44 #e-3
+                                                        
+                                            
+        elif self.config.softening_procedure_type is SofteningProcedureType.excess_lime:
+            @self.Constraint(doc = "Excess lime addition") 
+            def eq_excess_CaOH(b):
+                return b.excess_CaOH == (b.CO2_CaCO3 + prop_in.conc_mass_comp["Alkalinity_2-"] 
+                                         + b.Mg_hardness_CaCO3)*0.15
+        
+            @self.Constraint(doc= "Lime dosing")
+            def eq_CaOH_dosing(b):
+                return b.CaOH_dosing == (b.CO2_CaCO3 + prop_in.conc_mass_comp["Alkalinity_2-"] 
+                                         + b.Mg_hardness_CaCO3 + b.excess_CaOH)*prop_in.flow_vol * 0.56 #e-3
+            
+
+            @self.Constraint(doc="CO2 for first basin")
+            def eq_CO2_first_basin(b): 
+                return b.CO2_first_basin == (prop_in.conc_mass_comp["Alkalinity_2-"] - (b.Ca_CaCO3 + b.Ca_CaCO3)
+                                             + b.excess_CaOH  + prop_out.conc_mass_comp["Ca_2+"] * b.Ca_CaCO3_conv
+                                             + prop_out.conc_mass_comp["Mg_2+"]* b.Mg_CaCO3_conv) * prop_in.flow_vol * 0.44 #e-3
+
+
+        elif self.config.softening_procedure_type is SofteningProcedureType.single_stage_lime_soda:
+            @self.Constraint(doc= "Lime dosing")
+            def eq_CaOH_dosing(b):
+                return b.CaOH_dosing == pyunits.convert((b.CO2_CaCO3 + b.Ca_hardness_CaCO3) * 0.56 * prop_in.flow_vol,
+                                                         to_units=pyunits.kg/pyunits.d)
+            
+            @self.Constraint(doc = "Soda dosing")
+            def eq_Na2CO3_dosing(b):
+                return b.Na2CO3_dosing == pyunits.convert((b.Ca_hardness_nonCaCO3 + b.Mg_hardness_nonCaCO3 ) * 1.06 * prop_in.flow_vol,
+                                                         to_units=pyunits.kg/pyunits.d)
+            
+            @self.Constraint(doc="CO2 for first basin")
+            def eq_CO2_first_basin(b): 
+                return b.CO2_first_basin == pyunits.convert((prop_in.conc_mass_comp["Alkalinity_2-"] + 
+                                                            (b.Ca_hardness_nonCaCO3 + b.Mg_hardness_nonCaCO3 )- b.Ca_CaCO3 
+                                                + prop_out.conc_mass_comp["Ca_2+"])* b.Ca_CaCO3_conv * prop_in.flow_vol,to_units = pyunits.kg/pyunits.d) * 0.44 #e-3
+
+
+        else:
+            @self.Constraint(doc = "Excess lime addition") 
+            def eq_excess_CaOH(b):
+                return b.excess_CaOH == (b.CO2_CaCO3 + prop_in.conc_mass_comp["Alkalinity_2-"] 
+                                         + b.Mg_hardness_CaCO3)*0.15
+            
+
+            @self.Constraint(doc="Lime dosing")
+            def eq_CaOH_dosing(b):
+                return b.CaOH_dosing == pyunits.convert((b.CO2_CaCO3 + b.Ca_hardness_CaCO3 + 2*b.Mg_hardness_CaCO3 +
+                                                        b.Mg_hardness_nonCaCO3 + b.excess_CaOH) * 0.56 * prop_in.flow_vol,
+                                                         to_units=pyunits.kg/pyunits.d)
+
+            
+            def eq_Na2CO3_dosing(b):
+                return b.Na2CO3_dosing == pyunits.convert((b.Ca_hardness_nonCaCO3 + b.Mg_hardness_nonCaCO3 ) * 1.06 * prop_in.flow_vol,
+                                                         to_units=pyunits.kg/pyunits.d)
+            
+
+
+            @self.Constraint(doc="CO2 for first basin")
+            def eq_CO2_first_basin(b): 
+                return b.CO2_first_basin == pyunits.convert((b.excess_CaOH + prop_out.conc_mass_comp["Mg_2+"]* b.Mg_CaCO3_conv) * 
+                                                            prop_in.flow_vol,to_units = pyunits.kg/pyunits.d) * 0.44 #e-3
+            
+
+
+            @self.Constraint(doc="CO2 for second basin")
+            def eq_CO2_first_basin(b): 
+                return b.CO2_second_basin == pyunits.convert((prop_in.conc_mass_comp["Alkalinity_2-"] +
+                                                              (b.Ca_hardness_nonCaCO3 + b.Mg_hardness_nonCaCO3 ) - 
+                                                              (b.Ca_CaCO3 + b.Ca_CaCO3) +
+                                                              b.excess_CaOH + prop_out.conc_mass_comp["Mg_2+"]* b.Mg_CaCO3_conv) * 
+                                                            prop_in.flow_vol,to_units = pyunits.kg/pyunits.d) * 0.44 #e-3
+
+
+
+        if self.config.silica_removal:
+            if (value(prop_in.conc_mass_comp["SiO2"]) * 2.35 > value(prop_in.conc_mass_comp["Mg_2+"])) and value(prop_in.conc_mass_comp["SiO2"])> value( 0.050 * pyunits.kg / pyunits.m**3):
+                print('Mg dose')
+                @self.Constraint(doc="MgCl2 dosing constraint")
+                def eq_MgCl2_dosing(b):
+                    return b.MgCl2_dosing == (9.2 *prop_in.conc_mass_comp["SiO2"])
+        else:
+            pass
+
 
         @self.Constraint(doc="Recovery")
         def eq_recovery(b):
-            return prop_in.flow_vol_phase["Liq"] == prop_out.flow_vol_phase["Liq"] + prop_waste.flow_vol_phase["Liq"]
+            return prop_in.flow_vol == prop_out.flow_vol + prop_waste.flow_vol
 
         @self.Constraint(doc="Waste flow")
         def eq_waste_flow(b):
-            return prop_out.flow_vol_phase["Liq"] == prop_in.flow_vol_phase["Liq"] * b.frac_vol_recovery
+            return prop_out.flow_vol == prop_in.flow_vol * b.frac_vol_recovery
 
-        @self.Constraint(comps, doc="Mass balance")
+        @self.Constraint(non_hardness_comps, doc="Mass balance")
         def eq_mass_balance(b, j):
-            return prop_in.flow_mass_phase_comp[
-                "Liq", j
-            ] == prop_out.flow_mass_phase_comp["Liq", j] + prop_waste.flow_mass_phase_comp["Liq", j]
-
-        if "TSS" in comps:
-            @self.Constraint(doc="TSS removal")
-            def eq_effluent_tss(b):
-                return prop_out.conc_mass_phase_comp["Liq", "TSS"] == prop_in.conc_mass_phase_comp["Liq", "TSS"] * (1 - b.frac_TSS_removal)
-
-        @self.Constraint(doc="Ca in effluent")
-        def eq_effluent_ca(b):
-            return prop_out.conc_mass_caco3_comp["Ca_2+"] == b.ca_eff_target
-
-        if (
-            self.config.softening_procedure_type
-            is SofteningProcedureType.single_stage_lime
-        ):
-
-            if not self.config.silica_removal:
-
-                @self.Constraint(doc="Mg in effluent")
-                def eq_effluent_mg(b):
-                    return (
-                        prop_out.conc_mass_caco3_comp["Mg_2+"]
-                        == prop_in.conc_mass_caco3_comp["Mg_2+"]
-                    )
-
-            @self.Constraint(doc="Total hardness in effluent")
-            def eq_effluent_total_hardness(b):
-                return prop_out.total_hardness == (
-                    prop_out.conc_mass_caco3_comp["Ca_2+"]
-                    + (prop_in.alkalinity - prop_in.conc_mass_caco3_comp["Ca_2+"])
-                ) + (
-                    prop_in.total_hardness
-                    - (
-                        prop_in.conc_mass_caco3_comp["Ca_2+"]
-                        + (prop_in.alkalinity - prop_in.conc_mass_caco3_comp["Ca_2+"])
-                    )
-                )
-
-            @self.Constraint(doc="Alkalinity in effluent")
-            def eq_effluent_alkalinity(b):
-                return (
-                    prop_out.alkalinity
-                    == prop_in.alkalinity
-                    - prop_in.conc_mass_caco3_comp["Ca_2+"]
-                    # + b.ca_eff_target
-                    + prop_out.conc_mass_caco3_comp["Ca_2+"]
-                )
-
-            if "TDS" in comps:
-                if "SiO2" in comps:
-
-                    @self.Constraint(["TDS"], doc="TDS in effluent")
-                    def eq_effluent_tds(b, j):
-                        return prop_out.conc_mass_phase_comp[
-                            "Liq", "TDS"
-                        ] == prop_in.conc_mass_phase_comp["Liq", "TDS"] - (
-                            prop_in.conc_mass_phase_comp["Liq", "Ca_2+"]
-                            - prop_out.conc_mass_phase_comp["Liq", "Ca_2+"]
-                        ) - (
-                            prop_in.conc_mass_phase_comp["Liq", "SiO2"]
-                            - prop_out.conc_mass_phase_comp["Liq", "SiO2"]
-                        )
-
-                else:
-
-                    @self.Constraint(["TDS"], doc="TDS in effluent")
-                    def eq_effluent_tds(b, j):
-                        return prop_out.conc_mass_phase_comp[
-                            "Liq", "TDS"
-                        ] == prop_in.conc_mass_phase_comp["Liq", "TDS"] - (
-                            prop_in.conc_mass_phase_comp["Liq", "Ca_2+"]
-                            - prop_out.conc_mass_phase_comp["Liq", "Ca_2+"]
-                        )
+            return prop_in.flow_mass_comp[j] == prop_out.flow_mass_comp[j] + prop_waste.flow_mass_comp[j]
 
 
-
-            if "SiO2" in comps and self.config.silica_removal:
-                
-                self.mg_add_ratio = Param(
-                    initialize=9.1,
-                    units=pyunits.dimensionless,
-                    doc="Ratio of MgO: SiO2 to add for SiO2 removal"
-                )
-
-                self.mg_add = Var(
-                    initialize=10,
-                    units=pyunits.kg/pyunits.m**3,
-                    doc="Concentration MgO added for SiO2 removal"
-                )
-
-                @self.Constraint(doc="MgO addition")
-                def eq_mg_add(b):
-                    return b.mg_add == b.mg_add_ratio * prop_in.conc_mass_phase_comp["Liq", "SiO2"]
-
-                @self.Constraint(["Mg_2+"], doc="Mg in effluent")
-                def eq_effluent_mg(b, j):
-                    return (
-                        prop_out.conc_mass_caco3_comp["Mg_2+"]
-                        == prop_in.conc_mass_caco3_comp["Mg_2+"] + b.mg_add
-                    )
-
-
-
+        @self.Constraint(non_hardness_comps,doc="Component Removal")
+        def eq_component_removal(b,j):
+                return prop_waste.flow_mass_comp[j] == prop_in.flow_mass_comp[j] * (b.removal_efficiency[j])
+        
+        
         @self.Constraint(doc="Volume of mixer")
         def eq_volume_mixer(b):
             return (
                 b.volume_mixer
-                == b.properties_in[0].flow_vol_phase["Liq"] * b.retention_time_mixer
+                == pyunits.convert(b.properties_in[0].flow_vol * b.retention_time_mixer,to_units = pyunits.m**3) * b.no_of_mixer
             )
 
         @self.Constraint(doc="Volume of flocculator")
         def eq_volume_floc(b):
             return (
                 b.volume_floc
-                == b.properties_in[0].flow_vol_phase["Liq"] * b.retention_time_floc
+                == pyunits.convert( b.properties_in[0].flow_vol * b.retention_time_floc,to_units = pyunits.m**3) * b.no_of_floc
             )
 
         @self.Constraint(doc="Volume of sedimentation basin")
         def eq_volume_sed(b):
             return (
                 b.volume_sed
-                == b.properties_in[0].flow_vol_phase["Liq"] * b.retention_time_sed
+                == pyunits.convert(b.properties_in[0].flow_vol*b.retention_time_sed,to_units = pyunits.m**3)
             )
 
         @self.Constraint(doc="Volume of recarbonation basin")
         def eq_volume_recarb(b):
             return (
                 b.volume_recarb
-                == b.properties_in[0].flow_vol_phase["Liq"] * b.retention_time_recarb
+                == pyunits.convert(b.properties_in[0].flow_vol*b.retention_time_recarb, to_units = pyunits.m**3)
             )
 
-        # @self.Constraint(doc="Magnesium hydroxide acidity constant")
-        # def eq_acidity_const_mg_oh2(b):
-        #     return b.acidity_const_mg_oh2 == 4470.99 / b.properties_in[0].temperature + 0.01706 * b.properties_in[0].temperature - 6.0875
+        @self.Constraint(doc="Ca in effluent")
+        def eq_effluent_ca(b):
+            return prop_out.conc_mass_comp["Ca_2+"] == b.ca_eff_target / b.Ca_CaCO3_conv
+        
+        
+        @self.Constraint(doc="Mg in effluent")
+        def eq_effluent_mg(b):
+            return prop_out.conc_mass_comp["Mg_2+"] ==  b.mg_eff_target /  b.Mg_CaCO3_conv
+        
 
-        # @self.Constraint()
-        # def eq_excess_lime(b):
-        #     return b.excess_lime == (10 ** -b.properties_out[0].pOH / 2) * 74000
+        # The following conditionals check for softening methods (if excess lime term is required) and if silica removal is true (mgcl2 dosing term is required)
 
-        # @self.Constraint()
-        # def eq_pH_modified(b):
-        #     return b.properties_out[0].pH == log(10**-b.acidity_const_mg_oh2 / 10**-b.properties_out[0].pOH)
+        if (
+            self.config.softening_procedure_type is SofteningProcedureType.single_stage_lime or  self.config.softening_procedure_type is SofteningProcedureType.single_stage_lime_spda
+        ):
+
+            @self.Constraint(doc="Ca mass balance")
+            def eq_mass_balance_ca(b):
+                return (
+                        prop_waste.flow_mass_comp["Ca_2+"] ==  prop_in.flow_mass_comp["Ca_2+"] - prop_out.flow_mass_comp["Ca_2+"]
+                    )
+            
+            if self.config.silica_removal:
+                if ['TSS'] in comps:
+                    @self.Constraint(doc="Sludge production")
+                    def eq_sludge_prod(b):
+                        return b.sludge_prod == (2 * b.Ca_hardness_CaCO3 + 2.6 * b.Mg_hardness_CaCO3 + b.Ca_hardness_nonCaCO3 + 
+                            1.6*b.Mg_hardness_nonCaCO3  + prop_in.conc_mass_comp["TSS"] + b.MgCl2_dosing ) * prop_in.flow_vol
+                else:
+                    @self.Constraint(doc="Sludge production")
+                    def eq_sludge_prod(b):
+                        return b.sludge_prod == (2 * b.Ca_hardness_CaCO3 + 2.6* b.Mg_hardness_CaCO3 + b.Ca_hardness_nonCaCO3 + 
+                            1.6*b.Mg_hardness_nonCaCO3 + b.MgCl2_dosing) * prop_in.flow_vol
+
+                
+                @self.Constraint(doc="Mg mass balance")
+                def eq_mass_balance_mg(b):
+                    return (
+                            prop_waste.flow_mass_comp["Mg_2+"] ==  prop_in.flow_mass_comp["Mg_2+"] - prop_out.flow_mass_comp["Mg_2+"] +
+                            pyunits.convert(b.MgCl2_dosing * prop_in.flow_vol, to_units = pyunits.kg/pyunits.s)
+                            )         
+            
+                # if ["TDS"] in comps:
+                #     @self.Constraint(["TDS"], doc="TDS in effluent")
+                #     def eq_effluent_tds(b, j):
+                #         return prop_out.flow_mass_comp["TDS"] == ( prop_in.flow_mass_comp["TDS"] - (prop_in.flow_mass_comp[ "Ca_2+"]- prop_out.flow_mass_comp["Ca_2+"]) - 
+                #                                                 (prop_in.conc_mass_comp[ "SiO2"]- prop_out.conc_mass_comp["SiO2"]) )    
+
+            else : 
+                if ['TSS'] in comps:
+                    @self.Constraint(doc="Sludge production")
+                    def eq_sludge_prod(b):
+                        return b.sludge_prod == (2 * b.Ca_hardness_CaCO3 + 2.6 * b.Mg_hardness_CaCO3 + b.Ca_hardness_nonCaCO3 + 
+                            1.6*b.Mg_hardness_nonCaCO3  + prop_in.conc_mass_comp["TSS"]) * prop_in.flow_vol
+                else:
+                    @self.Constraint(doc="Sludge production")
+                    def eq_sludge_prod(b):
+                        return b.sludge_prod == (2 * b.Ca_hardness_CaCO3 + 2.6* b.Mg_hardness_CaCO3 + b.Ca_hardness_nonCaCO3 + 
+                            1.6*b.Mg_hardness_nonCaCO3  ) * prop_in.flow_vol
+
+                @self.Constraint(doc="Mg mass balance")
+                def eq_mass_balance_mg(b):
+                    return (prop_waste.flow_mass_comp["Mg_2+"] ==  prop_in.flow_mass_comp["Mg_2+"] - prop_out.flow_mass_comp["Mg_2+"])                  
+                
+                # if ["TDS"] in comps:
+                #     @self.Constraint(["TDS"], doc="TDS in effluent")
+                #     def eq_effluent_tds(b, j):
+                #         return prop_out.flow_mass_comp["TDS"] == prop_in.flow_mass_comp["TDS"] - (prop_in.flow_mass_comp[ "Ca_2+"]- prop_out.flow_mass_comp["Ca_2+"])
+
+        else:
+
+            @self.Constraint(doc="Ca mass balance")
+            def eq_mass_balance_ca(b):
+                return (
+                        prop_waste.flow_mass_comp["Ca_2+"] ==  prop_in.flow_mass_comp["Ca_2+"] - prop_out.flow_mass_comp["Ca_2+"] +
+                        pyunits.convert(b.excess_CaOH * prop_in.flow_vol , to_units =  pyunits.kg/pyunits.s)
+                    )
+        
+            if self.config.silica_removal:
+                if ['TSS'] in comps:
+                    @self.Constraint(doc="Sludge production")
+                    def eq_sludge_prod(b):
+                        return b.sludge_prod == (2 * b.Ca_hardness_CaCO3 + 2.6 * b.Mg_hardness_CaCO3 + b.Ca_hardness_nonCaCO3 + 
+                            1.6*b.Mg_hardness_nonCaCO3 + b.excess_CaOH + prop_in.conc_mass_comp["TSS"] + b.MgCl2_dosing ) * prop_in.flow_vol
+                else:
+                    @self.Constraint(doc="Sludge production")
+                    def eq_sludge_prod(b):
+                        return b.sludge_prod == (2 * b.Ca_hardness_CaCO3 + 2.6* b.Mg_hardness_CaCO3 + b.Ca_hardness_nonCaCO3 + 
+                            1.6*b.Mg_hardness_nonCaCO3 + b.excess_CaOH + b.MgCl2_dosing) * prop_in.flow_vol
+
+                
+                @self.Constraint(doc="Mg mass balance")
+                def eq_mass_balance_mg(b):
+                    return (
+                            prop_waste.flow_mass_comp["Mg_2+"] ==  prop_in.flow_mass_comp["Mg_2+"] - prop_out.flow_mass_comp["Mg_2+"] +
+                            pyunits.convert(b.MgCl2_dosing * prop_in.flow_vol, to_units = pyunits.kg/pyunits.s)
+                            )         
+            
+                # if ["TDS"] in comps:
+                #     @self.Constraint(["TDS"], doc="TDS in effluent")
+                #     def eq_effluent_tds(b, j):
+                #         return prop_out.flow_mass_comp["TDS"] == ( prop_in.flow_mass_comp["TDS"] - (prop_in.flow_mass_comp[ "Ca_2+"]- prop_out.flow_mass_comp["Ca_2+"]) - 
+                #                                                 (prop_in.conc_mass_comp[ "SiO2"]- prop_out.conc_mass_comp["SiO2"]) )    
+
+            else : 
+                if ['TSS'] in comps:
+                    @self.Constraint(doc="Sludge production")
+                    def eq_sludge_prod(b):
+                        return b.sludge_prod == (2 * b.Ca_hardness_CaCO3 + 2.6 * b.Mg_hardness_CaCO3 + b.Ca_hardness_nonCaCO3 + 
+                            1.6*b.Mg_hardness_nonCaCO3 + b.excess_CaOH  + prop_in.conc_mass_comp["TSS"]) * prop_in.flow_vol
+                else:
+                    @self.Constraint(doc="Sludge production")
+                    def eq_sludge_prod(b):
+                        return b.sludge_prod == (2 * b.Ca_hardness_CaCO3 + 2.6* b.Mg_hardness_CaCO3 + b.Ca_hardness_nonCaCO3 + 
+                            1.6*b.Mg_hardness_nonCaCO3  + b.excess_CaOH ) * prop_in.flow_vol
+
+                @self.Constraint(doc="Mg mass balance")
+                def eq_mass_balance_mg(b):
+                    return (prop_waste.flow_mass_comp["Mg_2+"] ==  prop_in.flow_mass_comp["Mg_2+"] - prop_out.flow_mass_comp["Mg_2+"])                  
+                
+                # if ["TDS"] in comps:
+                #     @self.Constraint(["TDS"], doc="TDS in effluent")
+                #     def eq_effluent_tds(b, j):
+                #         return prop_out.flow_mass_comp["TDS"] == prop_in.flow_mass_comp["TDS"] - (prop_in.flow_mass_comp[ "Ca_2+"]- prop_out.flow_mass_comp["Ca_2+"])
+    
+
 
     def initialize_build(
         blk,
@@ -489,6 +813,7 @@ class ChemicalSoftening0DData(InitializationMixin, UnitModelBlockData):
 
         state_args_waste = deepcopy(state_args)
 
+
         blk.properties_waste.initialize(
             outlvl=outlvl,
             optarg=optarg,
@@ -515,6 +840,30 @@ class ChemicalSoftening0DData(InitializationMixin, UnitModelBlockData):
 
     def calculate_scaling_factors(self):
         super().calculate_scaling_factors()
+
+        iscale.set_scaling_factor(self.retention_time_mixer, 1)
+        iscale.set_scaling_factor(self.retention_time_floc, 0.1)
+        iscale.set_scaling_factor(self.retention_time_sed, 1e-2)
+        iscale.set_scaling_factor(self.retention_time_recarb, 0.1)
+        iscale.set_scaling_factor(self.sedimentation_overflow, 0.1)
+        iscale.set_scaling_factor(self.no_of_mixer, 1)
+        iscale.set_scaling_factor(self.no_of_floc, 1)
+        iscale.set_scaling_factor(self.volume_mixer,0.1)
+        iscale.set_scaling_factor(self.volume_floc, 0.1)
+        iscale.set_scaling_factor(self.volume_sed, 0.1)
+        iscale.set_scaling_factor(self.volume_recarb, 0.1)
+        iscale.set_scaling_factor(self.vel_gradient_mix, 1e-2)
+        iscale.set_scaling_factor(self.vel_gradient_floc, 1e-2)
+        iscale.set_scaling_factor(self.frac_vol_recovery, 1)
+        iscale.set_scaling_factor(self.CaOH_dosing, 1e-5)
+        iscale.set_scaling_factor(self.Na2CO3_dosing, 1e-5)
+        iscale.set_scaling_factor(self.CO2_first_basin, 1e-2)
+        iscale.set_scaling_factor(self.CO2_second_basin, 1e-2)
+        iscale.set_scaling_factor(self.excess_CaOH, 1)
+        iscale.set_scaling_factor(self.CO2_CaCO3, 1e-2)
+        iscale.set_scaling_factor(self.MgCl2_dosing, 1e-2)
+        iscale.set_scaling_factor(self.sludge_prod, 1e-6)
+
 
     def _get_stream_table_contents(self, time_point=0):
         return create_stream_table_dataframe(
